# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Laporin API
  version: 1.0.0
  description: Backend API for Laporin internship report generator.
servers:
  - url: https://api.laporin.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  schemas:
...
    StudentInfo:
      type: object
      properties:
        full_name:
          type: string
        student_id:
          type: string
        school:
          type: string
        major:
          type: string
          nullable: true
        semester:
          type: string
          nullable: true
      required: [full_name, student_id, school]
    InternshipInfo:
      type: object
      properties:
        company_name:
          type: string
        company_address:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        role:
          type: string
          nullable: true
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
        company_supervisor:
          type: string
          nullable: true
        school_supervisor:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [company_name]
    ResearchSource:
      type: object
      properties:
        url:
          type: string
          format: uri
        title:
          type: string
          nullable: true
        fetched_at:
          type: string
          format: date-time
        confidence:
          type: number
          minimum: 0.0
          maximum: 1.0
          nullable: true
      required: [url, fetched_at]
    ResearchFact:
      type: object
      properties:
        claim:
          type: string
        sources:
          type: array
          items:
            $ref: '#/components/schemas/ResearchSource'
      required: [claim, sources]
    GeneratedSections:
      type: object
      properties:
        cover:
          type: string
        introduction:
          type: string
        company_profile:
          type: string
        activities:
          type: string
        conclusion:
          type: string
      required: [cover, introduction, company_profile, activities, conclusion]
...
    ReportUpdate:
      type: object
      properties:
        student:
          $ref: '#/components/schemas/StudentInfo'
        internship:
          $ref: '#/components/schemas/InternshipInfo'
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        status:
          type: string
          enum: [draft, researching, research_completed, generating, generated, preview_ready, payment_pending, paid, unlocked, failed, cancelled, expired]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        student:
          $ref: '#/components/schemas/StudentInfo'
        internship:
          $ref: '#/components/schemas/InternshipInfo'
        research_facts:
          type: array
          items:
            $ref: '#/components/schemas/ResearchFact'
        generated_sections:
          $ref: '#/components/schemas/GeneratedSections'
        generated_doc_path:
          type: string
          nullable: true
      required: [id, title, status, created_at]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
          required: [code, message]
      required: [error]
    GenericSuccess:
      type: object
      properties:
        message:
          type: string
      required: [message]
  responses:
    BadRequest:
      description: Invalid request payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    OK:
      description: Success
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GenericSuccess'
paths:
...
  /reports/{id}:
    get:
      summary: Retrieve a report project
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      summary: Update mutable fields of a report (e.g., student info)
      security:
        - sessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportUpdate'
      responses:
        '200':
          $ref: '#/components/responses/OK'
        '400':
          $ref: '#/components/responses/BadRequest'
...